+++ lib/src/utils/image_res.dart
@@ -1,0 +1,7 @@
+class ImagesRes {
+  /// Icons
+  static String info = "assets/icons/info.png";
+  static String error = "assets/icons/error.png";
+  static String success = "assets/icons/success.png";
+  static String warning = "assets/icons/warning.png";
+}